Moved unit test code to parent directory. Removed old test files. Migrated parse_args.py to use new sam_gen code
