--- test/test_positions.bed
@@ -1,20 +1,0 @@
-chr21	40769230	40769251	SRR2135332.47437/2	255	-
-chr4	80929693	80929715	SRR2135332.24366/1	255	-
-chr1	144915627	144916579	SRR2135332.34262/2	255	+
-chr6	170033046	170033067	SRR2135332.50858/1	255	-
-chr5	43175077	43175098	SRR2135332.37163/1	255	+
-chr1	144915627	144916579	SRR2135332.29340/2	255	+
-chr11	77632429	77632504	SRR2135332.48682/2	255	+
-chr11	2016691	2016713	SRR2135332.42392/2	255	+
-chr19	14071148	14071169	SRR2135332.43548/2	255	-
-chr21	40769230	40769251	SRR2135332.8403/1	255	+
-chr4	155490784	155490807	SRR2135332.43362/2	255	-
-chr19	54025328	54025881	SRR2135332.38765/2	255	-
-chr1	153233723	153233742	SRR2135332.46929/2	255	-
-chr3	53899230	53899305	SRR2135332.46243/1	255	+
-chr11	2016691	2016713	SRR2135332.42214/2	255	+
-chr18	25593677	25593698	SRR2135332.44303/2	255	+
-chr1	144915627	144916579	SRR2135332.6250/1	255	-
-chr7	94920428	94920449	SRR2135332.49197/2	255	-
-chr19	56243872	56243894	SRR2135332.13486/2	255	+
-chr1	156707541	156707562	SRR2135332.43716/1	255	-
